Refactor pages with styles to page-components
