+++ page-components/checkout/confirmation.js
@@ -1,0 +1,67 @@
+/* eslint react/no-multi-comp: 0, react/no-danger: 0 */
+import React from 'react';
+
+import { withTranslation } from 'lib/i18n';
+import { BasketContext } from 'components/basket';
+import Layout from 'components/layout';
+
+class Inner extends React.Component {
+  static contextType = BasketContext;
+
+  state = {
+    emptied: false
+  };
+
+  componentDidMount() {
+    this.empty();
+  }
+
+  empty() {
+    const { emptied } = this.state;
+    const { actions } = this.context;
+    if (!emptied) {
+      actions.empty();
+      this.setState({ emptied: true });
+    }
+  }
+
+  render() {
+    const { order, t } = this.props;
+    if (!order) {
+      return null;
+    }
+
+    return (
+      <div>
+        <h1>Thank you for the purchase!</h1>
+        <br />
+        <div>
+          Order id
+          {order.id}
+        </div>
+        <div>
+          Total price incl. tax:
+          {t('currency', {
+            amount: order.cart.total_price_including_tax / 100
+          })}
+        </div>
+      </div>
+    );
+  }
+}
+
+class CheckoutConfirmation extends React.Component {
+  static async getInitialProps(ctx) {
+    return { order: ctx.req.klarnaOrder };
+  }
+
+  render() {
+    return (
+      <Layout title="confirmation">
+        <Inner {...this.props} />
+      </Layout>
+    );
+  }
+}
+
+export default withTranslation(['common', 'checkout'])(CheckoutConfirmation);
