+++ page-components/folder/styles.js
