+++ pages/checkout/payment-gateway.js
@@ -1,39 +1,3 @@
-import React, { useState } from 'react';
-import styled from 'styled-components';
+import PaymentGatewayPage from 'page-components/checkout/payment-gateway';
 
-import { Button, Spinner } from 'ui';
-
-const Outer = styled.div`
-  flex: 0 1 500px;
-  background: #eee;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 1.5rem;
-  min-height: 500px;
-`;
-
-const InitiatingText = styled.div`
-  margin-right: 15px;
-`;
-
-const PaymentGateway = () => {
-  const [paying, setPaying] = useState(false);
-
-  return (
-    <Outer>
-      {paying ? (
-        <>
-          <InitiatingText>Initiating payment gateway...</InitiatingText>
-          <Spinner />
-        </>
-      ) : (
-        <Button large onClick={() => setPaying(true)}>
-          Pay now
-        </Button>
-      )}
-    </Outer>
-  );
-};
-
-export default PaymentGateway;
+export default PaymentGatewayPage;
